+++ group-13-project/src/app/query/route.ts
@@ -1,38 +1,5 @@
 
-import { sql } from '@vercel/postgres';
 
-import { UserProfile } from "@/lib/definitions";
-
-export async function fetchArtistById(){
-  const id = '891ccb08-e0b6-4aba-a439-33154b58dff0'
-  try {
-    const data = await sql<UserProfile>`
-      SELECT 
-        users.user_id,
-        users.name,
-        users.email,
-        users.image_url,
-        user_profiles.bio,
-        user_profiles.description,
-        user_profiles.artstyle,
-        user_profiles.instagram,
-        user_profiles.facebook,
-        user_profiles.pinterest
-      FROM users
-      LEFT JOIN user_profiles ON users.user_id = user_profiles.user_id
-      WHERE users.user_id = ${id};
-    `;
-    if (!data || data.rows.length === 0) {
-      return null; // No artist found, return null
-    }
-    const artist = data.rows[0];
-    console.log(artist);
-    return artist;
-  } catch (error) {
-    console.error('Database Error:', error);
-    throw new Error('Failed to fetch invoice.');
-  }
-}
 
 // async function listUserProfile() { 
 //     // Join `users` and `userprofiles` tables
@@ -59,7 +26,8 @@ export async function fetchArtistById(){
 export async function GET() {
 
   try {
-  	return Response.json(await fetchArtistById());
+    const artist = {};//await fetchArtistById();
+  	return Response.json(artist);
   } catch (error) {
   	return Response.json({ error }, { status: 500 });
   }
