+++ group-13-project/src/app/ui/skeletons.tsx
@@ -1,5 +1,4 @@
 import Image from "next/image"
-import Link from "next/link"
 
 export function GridSkeleton() {
   return (
