+++ group-13-project/src/lib/data.ts
@@ -48,6 +48,7 @@ export async function fetchUserProfiles() {
   `;
 
     return data.rows;
+    // eslint-disable-next-line
   } catch (error) {
     throw new Error("Failed to fetch user profiles.");
   }
